Separate styles out
