+++ src/components/Molecules/HeaderIcon/HeaderIcon.js
@@ -1,49 +1,18 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
 import Icon from '../../Atoms/SocialIcons/Icon/Icon';
 import HeaderIcons from './assets/HeaderIcons';
-import EmailSignUp from '../EmailSignUp/EmailSignUp';
 import RichText from '../../Atoms/RichText/RichText';
-import spacing from '../../../theme/shared/spacing';
 
-const IconWrapper = styled.div`
-  width: 35px;
-  height: 35px;
-  display: inline-block;
-  margin: 5px;
-`;
-
-const EsuWrapper = styled.div`
-  position: fixed;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(255, 255, 255, 0.5);
-  z-index: 100;
-  top: 0;
-  left: 0;
-  overflow-y: auto;
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-`;
-
-const CloseButton = styled(Icon)`
-  position: absolute;
-  top: 10px;
-  right: 25%;
-  width: 30px;
-  height: 30px;
-  margin: ${spacing('xl')} 10px 10px;
-`;
-
-const HeaderESU = styled(EmailSignUp)`
-  margin-top: ${spacing('xl')};
-  width: 50%;
-`;
+import {
+  IconWrapper,
+  EsuWrapper,
+  CloseButton,
+  HeaderESU
+} from './HeaderIcon.style';
 
 /* HeaderIcon component */
-const HeaderIcon = ({ link, campaign, iconType, title, ...restProps }) => {
+const HeaderIcon = ({ link, campaign, iconType, title }) => {
   const [isESUOpen, setIsESUOpen] = useState(false);
   const [isSuccess, setIsSuccess] = useState(false);
 
@@ -62,7 +31,7 @@ const HeaderIcon = ({ link, campaign, iconType, title, ...restProps }) => {
         <Icon
           onClick={e => handleESUClick(e, iconType)}
           icon={HeaderIcons[iconType]}
-          title="to-do"
+          title={title}
           brand={campaign}
           target="_blank"
           role="button"
@@ -90,7 +59,6 @@ const HeaderIcon = ({ link, campaign, iconType, title, ...restProps }) => {
             target="_self"
             role="button"
             href="#"
-            id="closey"
           />
         </EsuWrapper>
       ) : null}
